requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.14.0
httpx>=0.28.0
apscheduler>=3.10.0
jinja2>=3.1.0
pydantic-settings>=2.6.0
pyyaml>=6.0.0
python-multipart>=0.0.12
